AI/VectorIndexerWorker/pyproject.toml
[project]
name = "vectorindexerworker"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "confluent-kafka>=2.13.2",
    "grpcio>=1.62.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "qdrant-client>=1.17.0",
    "structlog>=25.5.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "testcontainers>=4.9",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
